+++ src/db/db.js
@@ -8,6 +8,17 @@ import eventBus from '../utils/broadcast-channel-pubsub'
 const SSPL_DB_VERSION = 2;
 export let db = null;
 
+import cacheRepository from './repository/cache';
+import groupsRepository from './repository/groups';
+import keyValueRepository from './repository/key-value';
+import playersRepository from './repository/players';
+import playersHistoryRepository from './repository/players-history';
+import rankedsRepository from './repository/rankeds';
+import rankedsChangesRepository from './repository/rankeds-changes';
+import scoresRepository from './repository/scores';
+import songsRepository from './repository/songs';
+import twitchRepository from './repository/twitch';
+
 export default async () => {
   IDBKeyRange.prototype.toString = function () {
     return "IDBKeyRange-" + (isDateObject(this.lower) ? this.lower.getTime() : this.lower) + '-' + (isDateObject(this.upper) ? this.upper : this.upper);
@@ -156,6 +167,9 @@ async function openDatabase(cache = null) {
       log.info("Create country ranks cache");
       await updateSongCountryRanks();
     }
+
+    // initialize all repositories in order to create cache to sync
+    [cacheRepository, groupsRepository, keyValueRepository, playersRepository, playersHistoryRepository, rankedsRepository, rankedsChangesRepository, scoresRepository, songsRepository, twitchRepository].map(repository => repository());
 
     return db;
   }