+++ src/db/repository/rankeds-changes.js
@@ -2,4 +2,4 @@ import createRepository from './generic';
 
 let repository;
 
-export default () => repository ? repository : repository = createRepository('rankeds-changes', '_idbId');
+export default () => repository ? repository : repository = createRepository('rankeds-changes', '_idbId', {'rankeds-changes-timestamp': 'timestamp', 'rankeds-changes-leaderboardId': 'leaderboardId'});