+++ src/db/repository/songs.js
@@ -2,4 +2,4 @@ import createRepository from './generic';
 
 let repository;
 
-export default () => repository ? repository : repository = createRepository('songs', 'hash');
+export default () => repository ? repository : repository = createRepository('songs', 'hash', {'songs-key': 'key'});