+++ src/scoresaber/rankeds.js
@@ -2,21 +2,18 @@ import rankedsRepository from "../db/repository/rankeds";
 import keyValueRepository from "../db/repository/key-value";
 import rankedsChangesRepository from "../db/repository/rankeds-changes";
 import {convertArrayToObjectByKey} from '../utils/js'
-import {getAccFromScore, getMaxScore, getMaxScoreFromSongCharacteristics} from '../song'
+import {getAccFromScore, getMaxScore} from '../song'
 import songsRepository from '../db/repository/songs'
 
 const RANKEDS_NOTES_CACHE_KEY = 'rankedsNotes';
 
-export const flushRankedsCache = () => rankedsRepository().flushCache();
-export const flushRankedsChangesCache = () => rankedsChangesRepository().flushCache();
 export const storeRanked = async ranked => rankedsRepository().set(ranked);
 export const storeRankeds = async rankeds => Promise.all(rankeds.map(async ranked => storeRanked(ranked)));
-export const getRankedSongs = async (refreshCache = false) => convertArrayToObjectByKey(await rankedsRepository().getAll(undefined, refreshCache) ?? {}, 'leaderboardId');
-export const getRankedSongsLastUpdated = async (refreshCache = true) => keyValueRepository().get('rankedSongsLastUpdated', refreshCache);
+export const getRankedSongs = async () => convertArrayToObjectByKey(await rankedsRepository().getAll() ?? {}, 'leaderboardId');
+export const getRankedSongsLastUpdated = async () => keyValueRepository().get('rankedSongsLastUpdated');
 export const setRankedSongsLastUpdated = async date => keyValueRepository().set(date,'rankedSongsLastUpdated');
-export const getRankedChanges = async query => rankedsChangesRepository().getAllFromIndex('rankeds-changes-timestamp', query);
 export const getRankedsChangesSince = async sinceTimestamp => {
-    const changes = await getRankedChanges(IDBKeyRange.lowerBound(sinceTimestamp));
+    const changes = await rankedsChangesRepository().getAllFromIndex('rankeds-changes-timestamp', IDBKeyRange.lowerBound(sinceTimestamp));
     if (!changes || !changes.length) return {};
 
     // return all song changes for matched timestamps {leaderboardId: [{change1}, {change2}...]}
@@ -45,7 +42,7 @@ export const setRankedsNotesCache = async rankedsNotesCache => keyValueRepositor
 
 export const getRankedsNotesCache = async () => {
     // try to get current cache
-    const currentCache = await keyValueRepository().get(RANKEDS_NOTES_CACHE_KEY, true);
+    const currentCache = await keyValueRepository().get(RANKEDS_NOTES_CACHE_KEY);
     if (currentCache) return currentCache;
 
     // prepare cache