+++ src/song.js
@@ -10,7 +10,7 @@ import {
 } from "./scoresaber/players";
 import {getActiveCountry} from "./scoresaber/country";
 import scoresRepository from './db/repository/scores'
-import {flushSsplCountryRanksCache, getSsplCountryRanks, setSsplCountryRanks} from './scoresaber/sspl-cache'
+import {getSsplCountryRanks, setSsplCountryRanks} from './scoresaber/sspl-cache'
 import eventBus from './utils/broadcast-channel-pubsub'
 import nodeSync from './network/multinode-sync'
 
@@ -91,7 +91,9 @@ export function findDiffInfo(characteristics, ssDiff) {
 
 export const updateSongCountryRanks = async (onlyLeaderboardsIds = null) => {
   const country = await getActiveCountry();
-  const countryPlayersIds = await getActiveCountryPlayersIds(country, true, true);
+  if (!country) return {};
+
+  const countryPlayersIds = await getActiveCountryPlayersIds(country, true);
 
   const ssplCountryRanks = await getSsplCountryRanks();
   const shouldProcessAllLeaderboards = isEmpty(ssplCountryRanks) || !onlyLeaderboardsIds;
@@ -118,17 +120,14 @@ export const updateSongCountryRanks = async (onlyLeaderboardsIds = null) => {
     ;
   });
 
-
   await setSsplCountryRanks(ssplCountryRanks);
-
-  flushSsplCountryRanksCache();
 
   eventBus.publish('sspl-country-ranks-cache-updated', {nodeId: nodeSync.getId(), ssplCountryRanks});
 
   return ssplCountryRanks;
 }
 
-export const getSongScores = async (leaderboardId, count = undefined, refreshCache = false) => scoresRepository().getAllFromIndex('scores-leaderboardId', leaderboardId, count, refreshCache);
+export const getSongScores = async (leaderboardId) => scoresRepository().getAllFromIndex('scores-leaderboardId', leaderboardId);
 
 export function getAccFromScore(score, maxSongScore) {
     const scoreMult = score.uScore && score.score ? score.score / score.uScore : 1
@@ -139,28 +138,28 @@ export function getAccFromScore(score, maxSongScore) {
         ? score.score / score.maxScoreEx / scoreMult * 100
         : null)
 }
-export async function getLeaderboard(leaderboardId, country, type = 'country', refreshCache = false) {
+export async function getLeaderboard(leaderboardId, country, type = 'country') {
     let players;
     switch(type) {
       case 'all':
-        players = await getAllActivePlayers(country, refreshCache);
+        players = await getAllActivePlayers(country);
         break;
 
       case 'manually_added':
-        players = await getFriends(country, true, refreshCache);
+        players = await getFriends(country, true);
         break;
 
       case 'country':
       default:
-        players = await getActiveCountryPlayers(country, true, refreshCache);
+        players = await getActiveCountryPlayers(country, true);
         break;
     }
     if (!players || !players.length) return [];
 
-    const songScores = await getSongScores(leaderboardId, undefined, refreshCache);
+    const songScores = await getSongScores(leaderboardId);
     if (!songScores) return [];
 
-    const playersIds = players.map(player => player.id);
+    const playersIds = players.filter(player => player.name).map(player => player.id);
 
     const filteredScores = songScores.filter(s => playersIds.includes(s.playerId));
     if (!filteredScores.length) return [];